+++ src/relstorage/tests/__init__.py
@@ -15,3 +15,42 @@ class TestCase(unittest.TestCase):
         self.assertEqual(len(container), 0)
 
     assertEmpty = assertIsEmpty
+
+class MockConnection(object):
+    rolled_back = False
+    closed = False
+    replica = None
+    committed = False
+
+    def commit(self):
+        self.committed = True
+
+    def rollback(self):
+        self.rolled_back = True
+
+    def close(self):
+        self.closed = True
+
+    def cursor(self):
+        return MockCursor()
+
+class MockCursor(object):
+    closed = False
+
+    def __init__(self):
+        self.executed = []
+        self.inputsizes = {}
+        self.results = []
+
+    def setinputsizes(self, **kw):
+        self.inputsizes.update(kw)
+
+    def execute(self, stmt, params=None):
+        params = tuple(params) if isinstance(params, list) else params
+        self.executed.append((stmt, params))
+
+    def fetchone(self):
+        return self.results.pop(0)
+
+    def close(self):
+        self.closed = True